+++ package.json
@@ -33,20 +33,22 @@
     "IPFS",
     "libp2p",
     "crypto",
-    "rsa"
+    "rsa",
+    "secp256k1"
   ],
   "license": "MIT",
   "dependencies": {
     "buffer": "^5.5.0",
     "err-code": "^2.0.0",
+    "is-typedarray": "^1.0.0",
     "iso-random-stream": "^1.1.0",
     "keypair": "^1.0.1",
-    "libp2p-crypto-secp256k1": "^0.4.2",
     "multibase": "^0.7.0",
     "multihashing-async": "^0.8.1",
     "node-forge": "~0.9.1",
     "pem-jwk": "^2.0.0",
     "protons": "^1.0.1",
+    "secp256k1": "^4.0.0",
     "ursa-optional": "~0.10.1"
   },
   "devDependencies": {
