--- secp256k1/package.json
@@ -1,66 +1,0 @@
-{
-  "name": "libp2p-crypto-secp256k1",
-  "version": "0.4.3",
-  "description": "Support for secp256k1 keys in libp2p-crypto",
-  "leadMaintainer": "Friedel Ziegelmayer <dignifiedquire@gmail.com>",
-  "main": "src/index.js",
-  "scripts": {
-    "lint": "aegir lint",
-    "build": "aegir build",
-    "test": "aegir test -t node -t browser",
-    "test:node": "aegir test -t node",
-    "test:browser": "aegir test -t browser",
-    "release": "aegir release",
-    "release-minor": "aegir release --type minor",
-    "release-major": "aegir release --type major",
-    "coverage": "aegir coverage",
-    "coverage-publish": "aegir coverage publish"
-  },
-  "keywords": [
-    "IPFS",
-    "libp2p",
-    "crypto",
-    "secp256k1"
-  ],
-  "license": "MIT",
-  "dependencies": {
-    "buffer": "^5.5.0",
-    "is-typedarray": "^1.0.0",
-    "multibase": "^0.7.0",
-    "multihashing-async": "^0.8.1",
-    "secp256k1": "^4.0.0"
-  },
-  "devDependencies": {
-    "aegir": "^21.0.2",
-    "benchmark": "^2.1.4",
-    "chai": "^4.2.0",
-    "dirty-chai": "^2.0.1",
-    "libp2p-crypto": "~0.17.2",
-    "protons": "^1.1.0"
-  },
-  "engines": {
-    "node": ">=6.0.0",
-    "npm": ">=3.0.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/libp2p/js-libp2p-crypto-secp256k1.git"
-  },
-  "bugs": {
-    "url": "https://github.com/libp2p/js-libp2p-crypto-secp256k1/issues"
-  },
-  "homepage": "https://github.com/libp2p/js-libp2p-crypto-secp256k1",
-  "contributors": [
-    "David Dias <daviddias.p@gmail.com>",
-    "Jacob Heun <jacobheun@gmail.com>",
-    "Friedel Ziegelmayer <dignifiedquire@gmail.com>",
-    "Hugo Dias <hugomrdias@gmail.com>",
-    "ᴠɪᴄᴛᴏʀ ʙᴊᴇʟᴋʜᴏʟᴍ <victorbjelkholm@gmail.com>",
-    "Yusef Napora <yusef@napora.org>",
-    "Alan Shaw <alan.shaw@protocol.ai>",
-    "Alberto Elias <hi@albertoelias.me>",
-    "Alex Potsides <alex@achingbrain.net>",
-    "Arve Knudsen <arve.knudsen@gmail.com>",
-    "Vasco Santos <vasco.santos@ua.pt>"
-  ]
-}
