+++ src/keys/index.js
@@ -13,7 +13,7 @@ exports = module.exports
 const supportedKeys = {
   rsa: require('./rsa-class'),
   ed25519: require('./ed25519-class'),
-  secp256k1: require('libp2p-crypto-secp256k1')(keysPBM, require('../random-bytes'))
+  secp256k1: require('./secp256k1-class')(keysPBM, require('../random-bytes'))
 }
 
 exports.supportedKeys = supportedKeys
