+++ test/fixtures/go-key-secp256k1.js
